RISC-V: Support separate firmware and kernel payload
This patch adds support for passing both the "-bios" and "-kernel"
options at the same time.  With this patch users no longer need to build
the kernel into the same ELF as the firmware, which makes the
development loop quite a bit quicker.  The mechanism for providing the
addresses of the kernel image to the firmware are a copy of how the
kernel expects an initramfs to be passed.  Specifically QEMU adds the
following device tree nodes

    chosen {
        riscv,kernel-start = <0x00000000 0x80200000>;
        riscv,kernel-end = <0x00000000 0x80590634>;
    };

These attributes are added by QEMU and read by BBL when combining "-bios
<firmware-image>" and "-kernel <kernel-image>" options. e.g.

    $ qemu-system-riscv64 -machine virt -bios bbl -kernel vmlinux

This load method could also be supported by a first stage boot loader
that reads seperate firmware and kernel from SPI flash.  The main
advantage of this new mechanism is that it eases kernel development by
allowing users to skip re-building the firmware whenever the kernel is
updated.

The board is also changed to use the firmware address from the loaded
firmware or combined firmware+kernel. This is normally equal to the DRAM
base address of 0x8000_0000, however now it is possible to boot firmware
at different load addresses because the reset code jumps to the actual
firmware entry address.

Signed-off-by: Michael Clark <mjc@sifive.com>
Palmer: Forward-ported the patch, updated the commit message.
Signed-off-by: Palmer Dabbelt <palmer@sifive.com>